sendmail {mail}	R Documentation

Sends an Email with Status Information

Description

Use sendmail to send an email notification with status information from R.

Usage

sendmail(recipient, subject="Notification from R",
		message="Calculation finished!", password="rmail")

Arguments

recipient
A valid email address.

subject
Subject of the email message.

message
Body of the email message.

password
Password. Please don't change the default password.

Details

Only a combination of alphabetic and numeric characters with the following special characters are valid: 0-9, a-z, A-Z, . : , ; ! ? % _ = @ *. Supported white spaces are blancs and line breaks. Line breaks must be encoded via \n or \r.

Value

The function sendmail returns a string with information about the processing of your message.

Note

To avoid spam, you have a maximum of 20 email notifications per day, that you can send from the same IP-address. Send me a message, if you want to deactivate the limitation of 20 email notifications per day. Notifications via SMS are supported only on request.

Author(s)

Dr. Lin Himmelmann <Rmail@linhi.com>, Statistical Consulting

Examples

# sendmail("your@email.com", "R notice", "Calculation finished.\nFetch your data!")

[Package mail version 1.0 Index]